Change card test to use constructor and CreditCard objects
This patch simplified all the tests in detectNetwork(), from inline
tests against numbers to making each type of network it's own CreditCard
object with a check() function built in. detectNetwork.test.js was also
updated to have a more concise iteration function, drawing on a small
library of prefixes and filler suffixes.